Everything in new (and better) format now. BruteForce solver takes less than 2 minutes to solve all of the solveable.txt grids

+++ SuDoKuSolver/defines.hpp
@@ -6,7 +6,7 @@
 #include <cstdint>
 
 typedef int32_t INT;
-typedef uint32_t UINT;
+typedef size_t UINT;
 typedef float FLT;
 
 // Bounds checking only when debug mode
